append city result to view
